rename: auth 파일 구조 수정 - #15
