+++ src/main/java/com/example/swcompetitionproject/entity/User.java
@@ -15,21 +15,6 @@ import java.util.List;
 @Getter
 @Table(name = "users")
 public class User extends BaseEntity{
-    ////이름
-    //@Column(nullable = false)
-    //private String name;
-//
-    ////전공(소속)
-    //@Column(nullable = false)
-    //private String major;
-//
-    ////구분
-    //@Column(nullable = false)
-    //private String division;
-//
-    ////학년
-    //@Column(nullable = false)
-    //private int grade;
 
     //학번
     @Column(nullable = false)
